replace the router from pwa-helpers with Vaadin.Router
Vaadin.Router is a small and powerful router for Web Components. It enables a full set of features needed to create app-like experiences on the web: lazy loading, animated transitions, navigation guards, parent layouts, redirects and more. It is framework-agnostic and works equally well with all Web Components, so it would be a good fit for the PWA starter kit.

See https://github.com/vaadin/vaadin-router for details.